chore: rename packages and files for consistency and clarity
