--- MovieBookingSystem/src/Services/PaymentService/PaymentMethod.java
@@ -1,10 +1,0 @@
-package Services.PaymentService;
-import Models.Payment;
-
-interface PaymentMethod {
-
-    public boolean pay(Payment payment);
-
-
-
-}
