+++ MovieBookingSystem/src/services/PaymentService/Gpay.java
@@ -1,0 +1,15 @@
+package services.PaymentService;
+
+import models.Payment;
+
+public class Gpay implements PaymentMethod{
+
+
+    @Override
+    public boolean pay(Payment payment) {
+        System.out.println("Amount paid successfully  Rs: "+ payment.getAmount() +" PaymentMethod :  "+payment.getPaymentMethod());
+        return true;
+    }
+
+    
+}
