+++ MovieBookingSystem/src/services/PaymentService/PaymentMethod.java
@@ -1,0 +1,10 @@
+package services.PaymentService;
+import models.Payment;
+
+interface PaymentMethod {
+
+    public boolean pay(Payment payment);
+
+
+
+}
